ffb-revision: Use ForcePercent as gain; fix periodic; change defaults
Force percent is supposed to be for overall effects strength, so use it as gain instead of using motor strength as gain. Motor strength will now be clamped to the nominal force.

For periodic forces, the magnitude is used as peak-to-peak amplitude, so input forces below the midrange cause increasingly smaller waveform effects. By ensuring that the magnitude is at least half of nominal, the effects will be stronger.

Default periods for each motor is now 1000 ms for a more pronounced periodic effect.