Enabled pass through to work even in combined mode (for a single controller).
This change required moving a notable chunk of logic out of x360ce.cpp into virtual methods in ControllerBase.cpp. The code is largely unchanged, but because it now exists as virtual methods in the base class, derived classes can change the implementation. ControllerCombiner will change the behavior of many of these methods to handle pass through in a different way.

In the new implementation, while controllers are getting created in ControllerCombiner we check to see if any of them are set to pass through. If so, we hold onto that controller in a special place inside of ControllerCombiner. Later, methods like GetBatteryInformation will use the pass through controller if there is one, otherwise the call is passed to the base method (which works exactly like it does today).